Clarify purpose of testLogger
